Add truth-bounded dashboard decision signals

+++ apps/agentacct/Sources/agentacct/UsagePane.swift
@@ -588,7 +588,7 @@ struct UsagePeriodChart: View {
         case .cost:
             return period.costText == "—" ? "unpriced" : period.costText
         case .tokens:
-            if let tokens = value(period) { return UsageTotals.compact(Int(tokens)) }
+            if let tokens = value(period) { return UsageTotals.compact(tokens) }
             return "none recorded"
         }
     }
@@ -598,7 +598,7 @@ struct UsagePeriodChart: View {
         case .cost:
             return "$" + String(format: maxValue * fraction >= 10 ? "%.0f" : "%.2f", maxValue * fraction)
         case .tokens:
-            return UsageTotals.compact(Int(maxValue * fraction))
+            return UsageTotals.compact(maxValue * fraction)
         }
     }
